add global function to separate file
